+++ packages/ricos-content/web/src/normalization/normalizeInitialState.ts
@@ -1,4 +1,4 @@
-import { cloneDeep, mapValues } from 'lodash';
+import { cloneDeep, mapValues, isEmpty } from 'lodash';
 import { processContentState } from './processContentState';
 import {
   IMAGE_TYPE,
@@ -8,8 +8,7 @@ import {
   VIDEO_TYPE_LEGACY,
   IMAGE_TYPE_LEGACY,
   VERTICAL_EMBED_TYPE,
-  COLLAPSIBLE_LIST_TYPE,
-  TABLE_TYPE,
+  WRAP,
 } from '../consts';
 import {
   linkDataNormalizer,
@@ -17,8 +16,6 @@ import {
   galleryDataNormalizer,
   videoDataNormalizer,
   verticalEmbedDataNormalizer,
-  collapsibleListDataNormalizer,
-  tableDataNormalizer,
 } from './dataNormalizers';
 import { ComponentData, DraftContent, NormalizeConfig, RicosEntity } from '../types';
 
@@ -35,8 +32,6 @@ const dataNormalizers: {
   [GALLERY_TYPE]: galleryDataNormalizer,
   [VIDEO_TYPE]: videoDataNormalizer,
   [VERTICAL_EMBED_TYPE]: verticalEmbedDataNormalizer,
-  [COLLAPSIBLE_LIST_TYPE]: collapsibleListDataNormalizer,
-  [TABLE_TYPE]: tableDataNormalizer,
 };
 
 const normalizeComponentData = (
@@ -99,8 +94,6 @@ const entityTypeMap = {
     [GALLERY_TYPE]: GALLERY_TYPE,
     [VIDEO_TYPE]: VIDEO_TYPE,
     [VERTICAL_EMBED_TYPE]: VERTICAL_EMBED_TYPE,
-    [COLLAPSIBLE_LIST_TYPE]: COLLAPSIBLE_LIST_TYPE,
-    [TABLE_TYPE]: TABLE_TYPE,
   },
 };
 
@@ -137,6 +130,9 @@ const normalizeEntityMap = (
         data: normalizeComponentData(entity.type, entity.data, config, stateVersion),
       };
     }
+    if (newEntity?.data?.config && !newEntity?.data?.config?.textWrap) {
+      newEntity.data.config.textWrap = WRAP;
+    }
     convertAnchorToLinkToUndoOneAppFix(newEntity);
     return newEntity;
   });
@@ -165,10 +161,16 @@ const convertAnchorToLinkToUndoOneAppFix = (newEntity: RicosEntity) => {
 
 export default (content: DraftContent, config: NormalizeConfig = {}): DraftContent => {
   const { blocks, entityMap, docStyle, VERSION } = processContentState(cloneDeep(content), config);
-  return {
-    blocks,
-    entityMap: normalizeEntityMap(entityMap, config, content.VERSION || '0.0.0'),
-    docStyle,
-    VERSION,
-  };
+  return isEmpty(docStyle)
+    ? {
+        blocks,
+        entityMap: normalizeEntityMap(entityMap, config, content.VERSION || '0.0.0'),
+        VERSION,
+      }
+    : {
+        blocks,
+        entityMap: normalizeEntityMap(entityMap, config, content.VERSION || '0.0.0'),
+        docStyle,
+        VERSION,
+      };
 };